+++ src/de/metal_land/tsp/Route.java
@@ -21,16 +21,16 @@ public class Route {
         route.add(node);
     }
 
-    public void greedy(LinkedList<Node> nodes, Map<Node,Map<Node,Double>> distances){
+    public void greedy(LinkedList<Node> nodes, Map<Node,Map<Node,Integer>> distances){
         route.clear();
         route.add(nodes.getFirst());
         nodes.removeFirst();
 
         while(!nodes.isEmpty()){
             Node lastNode = nodes.get(nodes.size()-1);
-            Map<Node,Double> distanceMap = distances.get(lastNode);
+            Map<Node,Integer> distanceMap = distances.get(lastNode);
             Node shortestNode = null;
-            Double shortestDistance = Double.MAX_VALUE;
+            Integer shortestDistance = Integer.MAX_VALUE;
 
             for (Node node : distanceMap.keySet()) {
                 if(distanceMap.get(node) < shortestDistance) {
@@ -44,9 +44,9 @@ public class Route {
         }
     }
 
-    public double calculateDistance(){
+    public int calculateDistance(){
         Node lastNode = null;
-        double distance = 0;
+        int distance = 0;
 
         for (Node node : route) {
             if(lastNode == null){
